LPS-166413 Simplifying similar implementations in one single point at PortalInstances class
