+++ portal-impl/src/com/liferay/portal/util/packageinfo
@@ -1,1 +1,1 @@
-version 41.0.0
+version 41.1.0
